shooting complett but hits missing

+++ Gun.pde
@@ -1,7 +1,7 @@
 class Gun extends Weapon {
   
-  int bulletSpeedX = 5;
-  int bulletSpeedY = 5;
+  float bulletSpeedX = 5;
+  float bulletSpeedY = 5;
   float bulletPositionX = 5000;
   float bulletPositionY = 5000;
